Fix weird unbind behaviour
We should not detach a connection from the event loop. This connection
might not have been closed correctly. It could be still open, so
eventmachin would abort with an assertion error.

Instead we should close the connection via `close_connection`. This
will close it correctly and remove it from the reactor.